Moved VendorExtensions closer to Vendor query handlers using Resharper.
