+++ Application/EdFi.Ods.AdminApp.Management/Database/Queries/VendorExtensions.cs
@@ -1,13 +1,12 @@
-﻿// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: Apache-2.0
 // Licensed to the Ed-Fi Alliance under one or more agreements.
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
 
 using System.Linq;
 using EdFi.Admin.DataAccess.Models;
-using EdFi.Ods.AdminApp.Management;
 
-namespace EdFi.Ods.AdminApp.Web.Helpers
+namespace EdFi.Ods.AdminApp.Management.Database.Queries
 {
     public static class VendorExtensions
     {
