+++ Application/EdFi.Ods.AdminApp.Web/Models/ViewModels/Global/AddVendorModel.cs
@@ -4,6 +4,7 @@
 // See the LICENSE and NOTICES files in the project root for more information.
 
 using EdFi.Ods.AdminApp.Management.Database.Commands;
+using EdFi.Ods.AdminApp.Management.Database.Queries;
 using EdFi.Ods.AdminApp.Web.Helpers;
 using FluentValidation;
